+++ src/webappScripts/createWebappConfig.ts
@@ -6,7 +6,7 @@ import { glob } from '../utils/glob-promise';
 import { getStaticFileDirectories } from './getStaticFileDirectories';
 
 export async function createWebappConfig({argv, cwd, zeroconfigPath}: {argv: WebappArgv, cwd: string, zeroconfigPath: string}): Promise<WebappConfig> {
-  const {command, app, sizeReport, mode, appFileName, vendorFileName, styleFileName, chunkPath, publicPath, port, serverPort, https} = argv;
+  const {command, app, sizeReport, mode, appFileName, vendorFileName, styleFileName, chunkPath, publicPath, port, serverPort, https, internalEslint} = argv;
   
   if (!(await fs.pathExists(path.join(cwd, 'src', app)))) throw new Error(`${path.join(cwd, 'src', app)} is undefined`);
   
@@ -35,6 +35,7 @@ export async function createWebappConfig({argv, cwd, zeroconfigPath}: {argv: Web
     styleFileName,
     chunkPath,
     publicPath,
+    internalEslint,
     
     port,
     serverPort,